// _archive/LUNEO_V1_PERSONALISATION/packages/virtual-try-on/src/core/types.ts
export interface RenderOptions {
  /** Activer antialiasing */
  antialias?: boolean;
  
  /** Activer ombres */
  shadows?: boolean;
  
  /** Intensité lumière ambiante */
  ambientLightIntensity?: number;
}